requirements.txt
pandas
google-cloud-bigquery
pyarrow